+++ src/hooks/useDetailData.js
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
-import { instance } from '../api/axios';
-import { requests } from '../api/requests';
+import { instance } from '../services/api';
+import { requests } from '../services/requests';
 
 export const useDetailData = (isOpen, movieId, movieType) => {
   const [movie, setMovie] = useState(null);
